% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configs}
\alias{configs}
\title{Configurations}
\usage{
configs(path = Sys.getenv("HOME"), progress = "interactive")
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the `HOME` directory. This parameter is then passed in [list.dirs]}
}
\description{
Get a data frame of the different git configurations
}
