% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Contributions}
\usage{
contributions(path = Sys.getenv("HOME"), breaks = "day",
  by = "author", progress = "interactive")
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the `HOME` directory. This parameter is then passed in [list.dirs]}

\item{breaks}{Default is \code{month}. Change to year, quarter,
week or day as necessary.}

\item{by}{Contributions by "commits" or "author". Default is "commits".}
}
\description{
Get some statistics about working projects
}
