% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ggit}
\alias{ggit}
\alias{ggit-package}
\title{Provides visualization routines for Git logs}
\description{
This package is a helper package to the R implementation of Git,
git2r. It provides routines to visualize activity in git
repositories, such as commit frequency, author contributions and
also (eventually) allow you to generate a full contribution graph
(like you see on a GitHub profile) for one or more repositories.
}

