% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_git_repo.R
\name{list_git}
\alias{list_git}
\title{List the Git Repository in a Directory/Folder}
\usage{
list_git(path = Sys.getenv("HOME"))
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the `HOME` directory. This parameter is then passed in [list.dirs]}
}
\description{
This function produces a character vector of the names of directories where a `.git` folder can be found.
}
