% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{punch_card}
\alias{punch_card}
\alias{punch_card,git_repository-method}
\title{Punch card}
\usage{
punch_card(repo)

\S4method{punch_card}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository}
}
\value{
A \code{ggplot} chart
}
\description{
Punch card
}
\keyword{methods}

