% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_git.R
\name{tbl_git}
\alias{tbl_git}
\title{Build a Tibble of Git Repositories}
\usage{
tbl_git(path = Sys.getenv("HOME"))
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the `HOME` directory. This parameter is then passed in [list.dirs]}
}
\description{
This function produces a tibble based on git repositories.
}
